Added the irc_client class with an event system.
The irc_client is in the main thread at the moment. If decided,
it would be feasible and perhaps less complex to keep it this way.
The starting file is now mud_boy.py rather than main.py, and events
have been set up for ping, PM and channel messages, along with a
test !help command.